updating everything for git clone
